Add functionality to orient and move robot
This commit adds support for the "command" part of the user input. After
parsing the user commands, it reads through each and either:
- changes the robot orientation
- moves the robot forward

Finally, the position and direction of the robot are printed to the
output screen.

This does _not_ account for bad commands (i.e. robot is lost) yet.