Add check to see if robot goes off-grid
This commit adds functionality to check if a given "F" command sends the
robot off-grid. If so, this value is stored on the robot object, and the
corresponding grid square (i.e. the one before the edge) is marked with
a "scent" as well as which particular forward coordinate caused the
robot to drop off.